practice1/requirements.txt
fastapi
uvicorn
numpy
pywavelets
scipy
Pillow
matplotlib
ffmpeg-python
python-multipart

